Consolidate util.js
