--- apply.js
@@ -1,4 +1,0 @@
-export default function apply (signal, fn, ...params) {
-	for (let i = 0, N = signal.length; i < N; i++) signal[i] *= fn(i, N, ...params)
-	return signal
-}
