--- generate.js
@@ -1,5 +1,0 @@
-export default function generate (fn, N, ...params) {
-	let w = new Float64Array(N)
-	for (let i = 0; i < N; i++) w[i] = fn(i, N, ...params)
-	return w
-}
